patch float to int in G_RunThink
This fixes potential precision issues if the map remains running for a very long time.

+++ code/game/g_main.c
@@ -1329,7 +1329,7 @@ Runs thinking code for this frame if necessary
 =============
 */
 void G_RunThink (gentity_t *ent) {
-	float	thinktime;
+	int	thinktime;
 
 	thinktime = ent->nextthink;
 	if (thinktime <= 0) {
